Table Alias Calculator should be an interface
