+++ src/main/java/org/mybatis/dynamic/sql/render/GuaranteedTableAliasCalculator.java
@@ -1,5 +1,5 @@
 /*
- *    Copyright 2016-2021 the original author or authors.
+ *    Copyright 2016-2022 the original author or authors.
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -27,7 +27,7 @@ import org.mybatis.dynamic.sql.SqlTable;
  * @author Jeff Butler
  *
  */
-public class GuaranteedTableAliasCalculator extends TableAliasCalculator {
+public class GuaranteedTableAliasCalculator extends ExplicitTableAliasCalculator {
 
     private GuaranteedTableAliasCalculator(Map<SqlTable, String> aliases) {
         super(aliases);
